Attempted fix download errors 4, tested on Windows NT 26100.0 running stable for 3 hours

+++ src/knmidownloader/DownloaderClient.cs
@@ -57,11 +57,22 @@ namespace knmidownloader
                 Console.WriteLine($"\nSending {folderName}/{name}. The hash differs.\n");
                 if (MainClass.Bot != null)
                 {
-                    if (MainClass.Bot.IsReady)
+                    if (MainClass.Bot.IsReady && File.Exists(filepath))
                     {
                         await MainClass.Bot.PostMessage(file.ID, filepath, msg);
                         await Task.Delay(10000);
                         File.Delete($"{MainClass.CurrentDir}/downloads/{type}/{folderName}/{name}");
+                    }
+                    else
+                    {
+                        if (MainClass.Bot.IsReady)
+                        {
+                            await MainClass.Bot.PostSystemMessage(4, $"Bot not ready or file missing\n\n{filepath}");
+                        }
+                        else
+                        {
+                            Console.WriteLine("Please end me");
+                        }
                     }
                 }
                 else
